requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cloudinary==1.40.0
Django==5.0.6
pillow==10.3.0
python-slugify==8.0.4
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
gunicorn==21.2.0